fe: Extract ChatMessage component to a separate JSX file. Clean up ChatMessage component. Improve style of ChatMessage.
